Fixed getting resolvedMode in popup.jsx
