+++ src/components/PopUp/PopUp.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
+import { useColorScheme } from "@mui/material";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Grid from "@mui/material/Grid";
@@ -16,6 +17,7 @@ import { theme } from "../../theme";
 import LogoEuCom from "../NavBar/images/SVG/Navbar/ep-logo.svg?url";
 import LogoInVidWeverify from "../NavBar/images/SVG/Navbar/invid_weverify.svg?url";
 import LogoVeraBlack from "../NavBar/images/SVG/Navbar/vera-logo_black.svg?url";
+import LogoVeraWhite from "../NavBar/images/SVG/Navbar/vera-logo_white.svg?url";
 
 const navigator = window.browser ? window.browser : window.chrome;
 
@@ -98,6 +100,9 @@ const PopUp = () => {
     }
   }, []);
 
+  const { systemMode, mode } = useColorScheme();
+  const resolvedMode = systemMode || mode;
+
   return (
     <ThemeProvider theme={theme}>
       <div className={classes.popUp}>
@@ -120,7 +125,7 @@ const PopUp = () => {
               </Grid>
               <Grid size={{ xs: 6 }}>
                 <img
-                  src={LogoVeraBlack}
+                  src={resolvedMode === "light" ? LogoVeraBlack : LogoVeraWhite}
                   alt={"logo_vera"}
                   style={{ width: "100px" }}
                 />
@@ -144,7 +149,7 @@ const PopUp = () => {
               </Grid>
               <Grid size={{ xs: 5 }}>
                 <img
-                  src={LogoVeraBlack}
+                  src={resolvedMode === "light" ? LogoVeraBlack : LogoVeraWhite}
                   alt={"logo_vera"}
                   style={{ width: "100px" }}
                 />
